Cloud computing project/notifier/requirements.txt
aiohttp==3.8.6
grpcio==1.59.*
protobuf==4.24.*
aiofiles==23.2.1